コードレビュー

+++ app/controllers/items_controller.rb
@@ -23,8 +23,8 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-   if user_signed_in?
     item = Item.find(params[:id])
+   if user_signed_in?&&current_user.id == item.user.id
     item.destroy
     redirect_to root_path
    else
